qagent/backend/requirements.txt
requests
python-dotenv
pydantic
google-generativeai
google-api-core
PyPDF2
pyyaml